DatabaseProject/backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
Werkzeug==2.3.7
PyJWT==2.8.0
bcrypt==4.0.1
python-dotenv==1.0.0
email-validator==2.0.0
mysql-connector-python==8.1.0
redis==4.6.0
Pillow==10.0.1
requests==2.31.0
psycopg2-binary